requirements (1).txt
requests==2.31.0
pyyaml==6.0.1
beautifulsoup4==4.12.2
pyjwt==2.8.0
jinja2==3.1.2